Separate out canvas and data url logic
